update filtersearch function
